# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste prático - PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/produtos": {
            "get": {
                "tags": [
                    "produto"
                ],
                "summary": "Listar produtos",
                "description": "Retornar todos os produtos",
                "operationId": "listar_produtos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/produtos-listagem": {
            "get": {
                "tags": [
                    "produto"
                ],
                "summary": "Listar produtos sem paginação",
                "description": "Retornar todos os produtos",
                "operationId": "listar_produtos_sem_paginacao",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/produto/{id}": {
            "get": {
                "tags": [
                    "produto"
                ],
                "summary": "Detalhar produtos",
                "description": "Retornar todos os produtos",
                "operationId": "detalhar_produto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "produto"
                ],
                "summary": "Excluir cadastro do produto",
                "operationId": "delete_produto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/produto{id?}": {
            "post": {
                "tags": [
                    "produto"
                ],
                "summary": "Cadastrar/Alterar produto",
                "operationId": "new_produto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos para cadastro",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "descricao": {
                                        "type": "name",
                                        "format": "text",
                                        "example": "Amortecedor Dianteiro"
                                    },
                                    "valor_venda": {
                                        "type": "number",
                                        "format": "number",
                                        "example": 1
                                    },
                                    "estoque": {
                                        "type": "number",
                                        "format": "number",
                                        "example": 1
                                    },
                                    "imagens[]": {
                                        "type": "file",
                                        "format": "binary",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    },
                    "422": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/produto-excluir-imagem/{id}": {
            "delete": {
                "tags": [
                    "produto"
                ],
                "summary": "Excluir imagem do produto",
                "operationId": "excluir_imagem_produto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário . "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}